Adds refactoring of security config classes
