+++ application/src/main/java/org/ehrbase/application/config/security/CorsConfig.java
@@ -1,0 +1,59 @@
+/*
+ * Copyright (c) 2021 Vitasystems GmbH and Jake Smolka (Hannover Medical School).
+ *
+ * This file is part of project EHRbase
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.ehrbase.application.config.security;
+
+import java.util.Arrays;
+import java.util.Collections;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.cors.CorsConfiguration;
+import org.springframework.web.cors.CorsConfigurationSource;
+import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+
+@Configuration
+public class CorsConfig {
+  @Bean
+  CorsConfigurationSource corsConfigurationSource() {
+    CorsConfiguration configuration = new CorsConfiguration();
+
+    configuration.applyPermitDefaultValues();
+    // Allow all origins to access EHRbase
+    configuration.setAllowedOrigins(Collections.singletonList("*"));
+    // Allowed HTTP methods
+    configuration.setAllowedMethods(
+        Arrays.asList("GET", "POST", "PUT", "DELETE", "HEAD", "OPTIONS"));
+    // Allow credentials to be transmitted
+    configuration.setAllowCredentials(true);
+    // Exposed headers that can be read by clients. Includes also all safe-listed headers
+    // See https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Expose-Headers
+    configuration.setExposedHeaders(
+        Arrays.asList(
+            "Access-Control-Allow-Methods",
+            "Access-Control-Allow-Origin",
+            "ETag",
+            "Content-Type",
+            "Last-Modified",
+            "Location",
+            "WWW-Authenticate"));
+
+    UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+    // Apply for all paths
+    source.registerCorsConfiguration("/**", configuration);
+    return source;
+  }
+}
