+++ application/src/main/java/org/ehrbase/application/config/security/SecurityConfig.java
@@ -1,0 +1,47 @@
+/*
+ * Copyright (c) 2020 Vitasystems GmbH and Hannover Medical School.
+ *
+ * This file is part of project EHRbase
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.ehrbase.application.config.security;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+
+@ConditionalOnProperty(name = "security.authType", havingValue = "NONE", matchIfMissing = true)
+@Configuration
+@EnableWebSecurity
+public class SecurityConfig extends WebSecurityConfigurerAdapter {
+
+  private final Logger logger = LoggerFactory.getLogger(getClass());
+
+  public SecurityConfig() {
+  }
+
+  @Override
+  protected void configure(HttpSecurity http) throws Exception {
+    logger.warn("Authentication disabled!");
+    logger.warn(
+        "To enable security set security.authType to BASIC or OAUTH in yaml properties file.");
+    http.cors().and().csrf().disable().authorizeRequests().anyRequest().permitAll();
+
+  }
+}
